requirements.txt
git+https://github.com/gorialis/jishaku
git+https://github.com/Rapptz/discord-ext-menus
discord.py
isort
black
asyncpg
asyncpg-stubs
yt-dlp
python-dateutil
playwright
lxml
parsedatetime
lru-dict
Pillow
pandas
beautifulsoup4
psutil
cachetools